# employee/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost/employees/employee/public/",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/dept_emp": {
            "get": {
                "tags": [
                    "Employees Departments"
                ],
                "summary": "Get all employees departments",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employees Departments"
                ],
                "summary": "Add a new employee to a department",
                "operationId": "App\\Http\\Controllers\\DeptEmpController::add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee number": {
                                        "type": "integar"
                                    },
                                    "Department number": {
                                        "type": "integar"
                                    },
                                    "from_date": {
                                        "type": "date"
                                    },
                                    "to_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "emp_no": 1,
                                    "dept_no": 1,
                                    "from_date": "2020-10-10",
                                    "to_date": "2020-08-09"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/dept_emp/{id}": {
            "put": {
                "tags": [
                    "Employees Departments"
                ],
                "summary": "Update employee department",
                "operationId": "App\\Http\\Controllers\\DeptEmpController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Department Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee number": {
                                        "type": "integar"
                                    },
                                    "Department number": {
                                        "type": "integar"
                                    },
                                    "from_date": {
                                        "type": "date"
                                    },
                                    "to_date": {
                                        "type": "date"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "emp_no": 1,
                                    "dept_no": 1,
                                    "from_date": "2020-10-10",
                                    "to_date": "2020-08-09"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employees Departments"
                ],
                "summary": "Delete Dempartment Employee attachment",
                "operationId": "App\\Http\\Controllers\\DeptEmpController::delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Department Employee id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Could Not Find Resource",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    }
}